#227 - clean up code for the 'additionalProperties' code assist, customizations for the 'property'
