+++ com.reprezen.swagedit.core/src/com/reprezen/swagedit/core/schema/ComplexTypeDefinition.java
@@ -19,6 +19,7 @@ import java.util.List;
 import com.fasterxml.jackson.core.JsonPointer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.google.common.collect.Iterables;
 import com.reprezen.swagedit.core.model.AbstractNode;
 
 /**
@@ -138,5 +139,10 @@ public class ComplexTypeDefinition extends TypeDefinition {
         }
 
         return isValid;
+    }
+    
+    @Override
+    public String getLabel() {
+        return Iterables.getFirst(complexTypes, null).getLabel();
     }
 }
