+++ com.reprezen.swagedit.core/src/com/reprezen/swagedit/core/validation/MultipleSwaggerErrorBuilder.java
@@ -10,20 +10,15 @@
  *******************************************************************************/
 package com.reprezen.swagedit.core.validation;
 
-import static com.google.common.collect.Iterators.transform;
-
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.google.common.base.Function;
-import com.google.common.base.Joiner;
 import com.google.common.base.Strings;
+import com.reprezen.swagedit.core.schema.JsonSchemaUtils;
 import com.reprezen.swagedit.core.validation.SwaggerError.MultipleSwaggerError;
 
 public class MultipleSwaggerErrorBuilder {
@@ -102,45 +97,7 @@ public class MultipleSwaggerErrorBuilder {
         if (swaggerSchemaNode == null) {
             return location;
         }
-        return getHumanFriendlyText(swaggerSchemaNode, location);
-    }
-
-    public static String getHumanFriendlyText(JsonNode swaggerSchemaNode, final String defaultValue) {
-        JsonNode title = swaggerSchemaNode.get("title");
-        if (title != null) {
-            return title.asText();
-        }
-        // nested array
-        if (swaggerSchemaNode.get("items") != null) {
-            return getHumanFriendlyText(swaggerSchemaNode.get("items"), defaultValue);
-        }
-        // "$ref":"#/definitions/headerParameterSubSchema"
-        JsonNode ref = swaggerSchemaNode.get("$ref");
-        if (ref != null) {
-            return getLabelForRef(ref.asText());
-        }
-        // Auxiliary oneOf in "oneOf": [ { "$ref": "#/definitions/securityRequirement" }]
-        JsonNode oneOf = swaggerSchemaNode.get("oneOf");
-        if (oneOf != null) {
-            if (oneOf instanceof ArrayNode) {
-                ArrayNode arrayNode = (ArrayNode) oneOf;
-                if (arrayNode.size() > 0) {
-                    Iterator<String> labels = transform(arrayNode.elements(), new Function<JsonNode, String>() {
-
-                        @Override
-                        public String apply(JsonNode el) {
-                            return getHumanFriendlyText(el, defaultValue);
-                        }
-                    });
-                    return "[" + Joiner.on(", ").join(labels) + "]";
-                }
-            }
-        }
-        return defaultValue;
-    }
-
-    private static String getLabelForRef(String refValue) {
-        return refValue.substring(refValue.lastIndexOf("/") + 1);
+        return JsonSchemaUtils.getHumanFriendlyText(swaggerSchemaNode, location);
     }
 
 }
