+++ com.reprezen.swagedit.openapi3/src/com/reprezen/swagedit/openapi3/schema/schema_v3.json
@@ -712,6 +712,7 @@
     },
     "schema": {
       "type": "object",
+      "title": "schema",
       "description": "The Schema Object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. This object is an extended subset of the JSON Schema Specification Wright Draft 00.  For more information about the properties, see JSON Schema Core and JSON Schema Validation. Unless stated otherwise, the property definitions follow the JSON Schema.",
       "additionalProperties": false,
       "patternProperties": {
@@ -836,7 +837,8 @@
         "properties": {
           "type": "object",
           "additionalProperties": {
-            "$ref": "#/definitions/schemaOrReference"
+            "$ref": "#/definitions/schemaOrReference",
+            "title": "property"
           }
         },
         "additionalProperties": {
